requirements.txt
openpyxl
numpy
plotly
pandas
streamlit
Pillow
scikit-learn
beautifulsoup4
streamlit_autorefresh
